Removed unnecessary extra file
